chore(refactor): rename util package to common
rename util package to common
