+++ .github/workflows/build-image.yml
@@ -1,14 +1,15 @@
 name: Build image
 on: [push, pull_request]
 jobs:
-
   build:
     name: Build
     runs-on: ubuntu-18.04
     steps:
+      - name: Check out mayadata-io/cstorpoolauto repo
+        uses: actions/checkout@v2
 
-    - name: Check out mayadata-io/cstorpoolauto repo
-      uses: actions/checkout@v2
+      - name: Run unit test
+        run: make test
 
-    - name: Build docker image
-      run: make image
+      - name: Build docker image
+        run: make image
