+++ common/metac/hook.go
