+++ common/pointer/pointer.go
