+++ common/string/string_test.go
